init commit

+++ pkg/peers/peers.go
@@ -9,7 +9,7 @@ import (
 
 type Peer struct {
 	IP   net.IP
-	Port uint32
+	Port uint16
 }
 
 func Unmarshal(peersBin []byte) ([]Peer, error) {
@@ -24,7 +24,7 @@ func Unmarshal(peersBin []byte) ([]Peer, error) {
 	for i := 0; i < numPeers; i++ {
 		offset := i * peerSize
 		peers[i].IP = net.IP(peersBin[offset: offset + 4])
-		peers[i].Port = binary.BigEndian.Uint32([]byte(peersBin[offset + 4: offset + 6]))
+		peers[i].Port = binary.BigEndian.Uint16([]byte(peersBin[offset + 4: offset + 6]))
 	}
 
 	return peers, nil
